Change MSExport to MSCoreExport
To have precise control of what is imported & exported from .dll, I changed all MSExport reference in MSCore to MSCoreExport.
MSCore.h & MSCore_Private.h both defines MSCoreExport as MSExport or MSImport depending of the context.
Those changes allow the use of MSImport & MSExport in other Frameworks and prevent confusion about what is imported or exported.
This wasn't an issue for Mac or Linux targets because MSExport & MSImport have the same values which is not the case in Win32