Fix MSColor initialize make programs using MSFoundation crashes.
The +load method of MSColor trigger the execution of the +initialize of MSColor before any autorelease pool is created.
When an autoreleased object is created by a called method, the runtime can't find any autorelease pool and raise an error message.
In wo451, this result in a log message via NSLog which in this context makes the program crash.