Add %lld support to NSString initWithFormat for wo451
wo451 doesn't support %lld in his formatting methods (NSLog, initWithFormat, ...) but it has a bit of %lld support with %qd.
Arguments passed to the initWithFormat are 32bits aligned and despite the given specifier length (hh, ll, ...) it always move by chuck of 32bits.
So NSLog(@"%qd", myInt64Value) works as expected, but NSLog(@"%qd %qd", myInt64Value, myInt64Value) do not, because the second value is printed with the 32 last bits of the first value and with the 32 first bits of the expected value.
With the %.0s trick, it's possible to get back to normal : NSLog(@"%qd%.0s %qd%0.s", myInt64Value, myInt64Value) work as NSLog(@"%lld %lld", myInt64Value, myInt64Value).
The %.0s trick force the _NSvsnprinf implementation to read a chunk of 32 bits without printing anything.
To hide all that process to end users, the initWithFormat method of NSPlaceholderString is replaced at runtime by the polyfill.
It's NSPlaceholderString and not NSString, because it's an NSPlaceholderString that [NSString alloc] returns.

This polyfill is tricky and should be replaced by our own stable and efficient solution.