Compatibility fix : convert OPENSSL_PEM_read_bio_RSAPublicKey to OPENSSL_PEM_read_bio_RSA_PUBKEY
It appears that a lot of openssl applications (including openssl itself) use RSA_PUBKEY instead of RSAPublicKey functions.
This changes makes public key creation and management easier as tools are much more likely to output a public key in that format.
For information since a very recent openssl version the openssl rsa command allow to use -RSAPublicKey_out to output a public key that is accepted by the RSAPublicKey functions