API : Improve authentication api security
To improve the security of the authentication and to allow the evolution of security algorithms, the auth API has to changes.

For public/private key part, the change is very small :
 - the challenge is not encoded to base64 before encryption
 - the decrypted challenge has to be encoded to base64
Those changes improve security around the possibility to leak user existance.
That leak could have been fixed by changing the way the pseudo crypt challenge is generated, but even so, because the underlying data are in a know range (base64 chars), some cryptographic attacks could applies.

For the login/password part, changes are much larger :
 - the hashed password now contains 4 parts :
   - algorithm : the algorithm used to generate the hash,
   - hardness : an easy to change way to increase hash password strength,
   - salt : a random generated buffer prepended to the password before the hash generation (deny rainbow tables),
   - hash : the resulting hash algo^hardness(salt + password)
sha512(pass) is a very weak way to protect password. Computing all combination of sha512(pass) with pass up to 8 chars now takes less than a day.
This simple approach is also very vulnerable to rainbow table attacks.
The new approach allow to increase security over time, by following computing capabilities evolution and the discovery of algorithm vulnerabilities.