[fix] [MSFoundation] fix the diamond problem with MSArray and MSDictionary
If an MSArray or MSDictionary is mutable (forever or not) it behave like it extends the NSMutable class.
To do so, the MSArray and MSDictionary classes use the forward invocation mecanism to pick NSMutable implementation in case of missing implement in MS classes.
With the approach MSArray and MSDictionary can use categories of NSMutable classes and can be used in place of NSMutable instances.

The NSArray, NSDictionary implementation use the generic methods of array and dictionary.